refactor(PrimaryFeature.jsx): Rename FeaturesDesktop & FeaturesMobile to ContentTabsDesktop & CarouselMobile and refactors PrimaryFeatures.jsx code to be more readable and maintainable
