+++ src/components/icon/DeviceNotificationIcon.jsx
@@ -1,0 +1,15 @@
+export const DeviceNotificationIcon = (props) => (
+  <svg viewBox="0 0 32 32" aria-hidden="true" {...props}>
+    <circle cx={16} cy={16} r={16} fill="#A3A3A3" fillOpacity={0.2} />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M9 0a4 4 0 00-4 4v24a4 4 0 004 4h14a4 4 0 004-4V4a4 4 0 00-4-4H9zm0 2a2 2 0 00-2 2v24a2 2 0 002 2h14a2 2 0 002-2V4a2 2 0 00-2-2h-1.382a1 1 0 00-.894.553l-.448.894a1 1 0 01-.894.553h-6.764a1 1 0 01-.894-.553l-.448-.894A1 1 0 0010.382 2H9z"
+      fill="#A3A3A3"
+    />
+    <path
+      d="M9 8a2 2 0 012-2h10a2 2 0 012 2v2a2 2 0 01-2 2H11a2 2 0 01-2-2V8z"
+      fill="#737373"
+    />
+  </svg>
+);
