+++ src/components/icon/NetworkIcon.jsx
@@ -1,0 +1,11 @@
+export const NetworkIcon = (props) => (
+  <svg viewBox="0 0 24 24" aria-hidden="true" {...props}>
+    <circle cx={12} cy={12} r={12} fill="#A3A3A3" fillOpacity={0.2} />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M3.412,0.516 C5.085,0.79 6.141,1.933 6.227,3.636 C6.224,4.372 5.966,5.057 5.511,5.631 C5.428,5.736 5.323,5.82 5.228,5.914 L7.773,9.307 C8.24,8.984 8.811,8.892 9.364,8.864 L9.499,8.87 L10.364,5.411 C9.722,5.025 9.401,4.368 9.364,3.636 C9.364,2.482 10.3,1.545 11.455,1.545 C12.609,1.545 13.545,2.482 13.545,3.636 C13.545,4.791 12.609,5.727 11.455,5.727 L11.363,5.723 L10.519,9.099 C11.207,9.355 11.77,9.862 12.121,10.505 C12.211,10.668 12.266,10.848 12.338,11.019 L17.734,10.038 L17.727,9.909 C17.727,8.177 19.131,6.773 20.864,6.773 C22.596,6.773 24,8.177 24,9.909 C24,11.641 22.596,13.045 20.864,13.045 C19.552,13.029 18.464,12.252 17.96,11.059 L12.497,12.053 C12.49,12.627 12.326,13.192 12.004,13.669 L17.612,18.155 C18.185,17.527 18.991,17.269 19.818,17.227 C21.55,17.227 22.955,18.631 22.955,20.364 C22.955,22.096 21.55,23.5 19.818,23.5 C18.086,23.5 16.682,22.096 16.682,20.364 C16.702,19.971 16.69,19.884 16.823,19.431 C16.866,19.283 16.942,19.148 17.002,19.006 L11.31,14.453 C10.746,14.913 10.079,15.1 9.364,15.136 C8.653,15.113 7.846,14.898 7.327,14.379 L4.975,16.339 C5.127,16.613 5.166,16.92 5.182,17.227 C5.182,18.382 4.246,19.318 3.091,19.318 C1.936,19.318 1,18.382 1,17.227 C1,16.072 1.936,15.136 3.091,15.136 C3.59,15.146 3.899,15.261 4.306,15.536 L6.66,13.575 C6.345,13.111 6.255,12.546 6.227,12 C6.248,11.31 6.466,10.482 6.97,9.979 L4.355,6.492 C3.966,6.703 3.525,6.751 3.091,6.773 C1.359,6.773 -0.045,5.369 -0.045,3.636 C-0.045,1.904 1.359,0.5 3.091,0.5 L3.412,0.516 z"
+      fill="#A3A3A3"
+    />
+  </svg>
+);
