+++ src/components/utilities/appScreenConfig.jsx
@@ -1,0 +1,5 @@
+import { motion } from "framer-motion";
+import { AppScreen } from "@/components/AppScreen";
+
+export const MotionAppScreenHeader = motion(AppScreen.Header);
+export const MotionAppScreenBody = motion(AppScreen.Body);
